+++ src/syntax/inline/html.ts
@@ -8,7 +8,7 @@ import { squash } from './text';
 type SubParsers = [InlineParser] | [PlainTextParser];
 
 const syntax = /^(<([a-z]+)>)/i;
-const inlinetags = Object.freeze('small|cite|q|dfn|abbr|data|time|code|var|samp|kbd|sub|sup|u|mark|ruby|rt|rp|bdi|bdo|ins|del'.split('|'));
+const inlinetags = Object.freeze('small|cite|q|dfn|abbr|data|time|code|var|samp|kbd|sub|sup|u|mark|ruby|rt|rp|bdi|bdo|ins|del|wbr'.split('|'));
 assert(inlinetags.every(t => /[a-z]+/.test(t)));
 assert(inlinetags.every(t => ['script', 'style', 'link', 'a', 'img'].indexOf(t) === -1));
 assert(inlinetags.every(t => ['s', 'em', 'strong'].indexOf(t) === -1));
@@ -18,6 +18,7 @@ export const html: HTMLParser = function (source: string): Result<HTMLElement, S
   if (!source.startsWith('<') || source.startsWith('<<')) return;
   const [whole, opentag, tagname] = source.match(syntax) || ['', '', ''];
   if (!whole) return;
+  if (['wbr'].indexOf(tagname.toLowerCase()) !== -1) return [[document.createElement(tagname.toLowerCase())], source.slice(opentag.length)];
   if (inlinetags.indexOf(tagname.toLowerCase()) === -1) return;
   if (!cache.has(tagname)) {
     void cache.set(tagname, new RegExp(`^</${tagname}>`, 'i'));