+++ src/view/segment.test.ts
@@ -7,19 +7,34 @@ describe('Unit: view/segment', () => {
     });
 
     it('ab', () => {
-      assert.deepStrictEqual(segment('\n'), ['\n']);
-      assert.deepStrictEqual(segment('\n\n'), ['\n\n']);
-      assert.deepStrictEqual(segment('\n\n\n'), ['\n\n\n']);
-      assert.deepStrictEqual(segment('\n\n\n\n'), ['\n\n\n\n']);
-      assert.deepStrictEqual(segment(' '), [' ']);
-      assert.deepStrictEqual(segment(' \n'), [' \n']);
       assert.deepStrictEqual(segment('a'), ['a']);
       assert.deepStrictEqual(segment('a\n'), ['a\n']);
+      assert.deepStrictEqual(segment('a\n\n'), ['a\n\n']);
+      assert.deepStrictEqual(segment('a\n\n\n'), ['a\n\n', '\n']);
       assert.deepStrictEqual(segment('a\nb'), ['a\nb']);
       assert.deepStrictEqual(segment('a\nb\n'), ['a\nb\n']);
       assert.deepStrictEqual(segment('a\nb\n\n'), ['a\nb\n\n']);
       assert.deepStrictEqual(segment('a\nb\n\n\n'), ['a\nb\n\n', '\n']);
       assert.deepStrictEqual(segment('a\nb\n\nc\n\nd'), ['a\nb\n\n', 'c\n\n', 'd']);
+      assert.deepStrictEqual(segment('a '), ['a ']);
+      assert.deepStrictEqual(segment(' a'), [' a']);
+      assert.deepStrictEqual(segment(' a '), [' a ']);
+    });
+
+    it('newline', () => {
+      assert.deepStrictEqual(segment('\n'), ['\n']);
+      assert.deepStrictEqual(segment('\n\n'), ['\n\n']);
+      assert.deepStrictEqual(segment('\n\n\n'), ['\n\n\n']);
+      assert.deepStrictEqual(segment('\n\n\n\n'), ['\n\n\n\n']);
+      assert.deepStrictEqual(segment(' '), [' ']);
+      assert.deepStrictEqual(segment(' \n \n \n '), [' \n \n \n', ' ']);
+      assert.deepStrictEqual(segment('a\n'), ['a\n']);
+      assert.deepStrictEqual(segment('a\n '), ['a\n ']);
+      assert.deepStrictEqual(segment('a\n\n '), ['a\n\n', ' ']);
+      assert.deepStrictEqual(segment('a\n\n\n '), ['a\n\n', '\n', ' ']);
+      assert.deepStrictEqual(segment('a\n\n\n\n '), ['a\n\n', '\n\n', ' ']);
+      assert.deepStrictEqual(segment('a\n\n\n\n\n '), ['a\n\n', '\n\n\n', ' ']);
+      assert.deepStrictEqual(segment('a\n\n\n\n\n\n '), ['a\n\n', '\n\n\n\n', ' ']);
     });
 
     it('pretext', () => {
@@ -29,6 +44,9 @@ describe('Unit: view/segment', () => {
       assert.deepStrictEqual(segment('````\n```\n````'), ['````\n```\n````']);
       assert.deepStrictEqual(segment('a\n\n\n```\n\n\n```'), ['a\n\n', '\n', '```\n\n\n```']);
       assert.deepStrictEqual(segment('a\n\n\n\n```\n\n\n```'), ['a\n\n', '\n\n', '```\n\n\n```']);
+      assert.deepStrictEqual(segment('a\n```\n```'), ['a\n```\n```']);
+      assert.deepStrictEqual(segment(' \n```'), [' \n', '```']);
+      assert.deepStrictEqual(segment('a\n```'), ['a\n```']);
     });
 
     it('extension', () => {
@@ -38,11 +56,17 @@ describe('Unit: view/segment', () => {
       assert.deepStrictEqual(segment('~~~~\n~~~\n~~~~'), ['~~~~\n~~~\n~~~~']);
       assert.deepStrictEqual(segment('a\n\n\n~~~\n\n\n~~~'), ['a\n\n', '\n', '~~~\n\n\n~~~']);
       assert.deepStrictEqual(segment('a\n\n\n\n~~~\n\n\n~~~'), ['a\n\n', '\n\n', '~~~\n\n\n~~~']);
+      assert.deepStrictEqual(segment(' \n~~~'), [' \n', '~~~']);
+      assert.deepStrictEqual(segment('a\n~~~'), ['a\n~~~']);
     });
 
     it('mixed', () => {
-      assert.deepStrictEqual(segment('a\n```'), ['a\n', '```']);
-      assert.deepStrictEqual(segment('a\n~~~'), ['a\n', '~~~']);
+      assert.deepStrictEqual(segment(' ```\n\n\n~~~\n\n\n```'), [' ```\n\n', '\n', '~~~\n\n', '\n', '```']);
+      assert.deepStrictEqual(segment(' ~~~\n\n\n```\n\n\n~~~'), [' ~~~\n\n', '\n', '```\n\n', '\n', '~~~']);
+      assert.deepStrictEqual(segment('a```\n\n\n~~~\n\n\n```'), ['a```\n\n', '\n', '~~~\n\n', '\n', '```']);
+      assert.deepStrictEqual(segment('a~~~\n\n\n```\n\n\n~~~'), ['a~~~\n\n', '\n', '```\n\n', '\n', '~~~']);
+      assert.deepStrictEqual(segment('```\n\n\n'), ['```\n\n', '\n']);
+      assert.deepStrictEqual(segment('~~~\n\n\n'), ['~~~\n\n', '\n']);
       assert.deepStrictEqual(segment('```\n\n\n~~~\n\n\n```'), ['```\n\n\n~~~\n\n\n```']);
       assert.deepStrictEqual(segment('~~~\n\n\n```\n\n\n~~~'), ['~~~\n\n\n```\n\n\n~~~']);
     });